+++ spec/jobs/create_manifest_job_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 require 'rails_helper'
 
-RSpec.describe CreateManifest, :clean do
+RSpec.describe CreateManifestJob, :clean do
   let(:generic_work)  { FactoryBot.create(:public_work, id: '888888', title: ['foo']) }
   let(:solr_document) { SolrDocument.new(attributes) }
   let(:request) { ENV['HOSTNAME'] = 'example.org' }
@@ -21,13 +21,14 @@ RSpec.describe CreateManifest, :clean do
   before do
     ENV['IIIF_MANIFEST_CACHE'] = "./tmp"
     allow(CurateGenericWork).to receive(:all).and_return([generic_work])
+    allow(CurateGenericWork).to receive(:find).and_return(generic_work)
     allow(SolrDocument).to receive(:find).and_return(solr_document)
     FileUtils.rm_f("./tmp/2019-11-11_18-20-32_888888")
   end
 
   it 'creates manifest and saves to tmp' do
     expect(File).not_to exist("./tmp/2019-11-11_18-20-32_888888")
-    described_class.process
+    described_class.perform_now
     expect(File).to exist("./tmp/2019-11-11_18-20-32_888888")
   end
 end
